+++ packages/api/src/router/widgets/index.ts
@@ -10,10 +10,12 @@ import { mediaRequestsRouter } from "./media-requests";
 import { mediaServerRouter } from "./media-server";
 import { mediaTranscodingRouter } from "./media-transcoding";
 import { minecraftRouter } from "./minecraft";
+import { networkControllerRouter } from "./network-controller";
 import { notebookRouter } from "./notebook";
 import { optionsRouter } from "./options";
 import { rssFeedRouter } from "./rssFeed";
 import { smartHomeRouter } from "./smart-home";
+import { stockPriceRouter } from "./stocks";
 import { weatherRouter } from "./weather";
 
 export const widgetRouter = createTRPCRouter({
@@ -22,6 +24,7 @@ export const widgetRouter = createTRPCRouter({
   app: appRouter,
   dnsHole: dnsHoleRouter,
   smartHome: smartHomeRouter,
+  stockPrice: stockPriceRouter,
   mediaServer: mediaServerRouter,
   calendar: calendarRouter,
   downloads: downloadsRouter,
@@ -33,4 +36,5 @@ export const widgetRouter = createTRPCRouter({
   minecraft: minecraftRouter,
   docker: dockerRouter,
   options: optionsRouter,
+  networkController: networkControllerRouter,
 });