+++ packages/forms-collection/src/icon-picker/icon-picker.tsx
@@ -33,9 +33,22 @@ interface IconPickerProps {
   error?: string | null;
   onFocus?: FocusEventHandler;
   onBlur?: FocusEventHandler;
+
+  label?: string;
+  placeholder?: string;
+  withAsterisk?: boolean;
 }
 
-export const IconPicker = ({ value: propsValue, onChange, error, onFocus, onBlur }: IconPickerProps) => {
+export const IconPicker = ({
+  value: propsValue,
+  onChange,
+  error,
+  onFocus,
+  onBlur,
+  withAsterisk = true,
+  label,
+  placeholder,
+}: IconPickerProps) => {
   const [value, setValue] = useUncontrolled({
     value: propsValue,
     onChange,
@@ -145,10 +158,10 @@ export const IconPicker = ({ value: propsValue, onChange, error, onFocus, onBlur
               setSearch(value || "");
             }}
             rightSectionPointerEvents="none"
-            withAsterisk
+            withAsterisk={withAsterisk}
             error={error}
-            label={tCommon("iconPicker.label")}
-            placeholder={tCommon("iconPicker.header", { countIcons: String(data?.countIcons ?? 0) })}
+            label={label ?? tCommon("iconPicker.label")}
+            placeholder={placeholder ?? tCommon("iconPicker.header", { countIcons: String(data?.countIcons ?? 0) })}
           />
           {session?.user.permissions.includes("media-upload") && (
             <UploadMedia