+++ packages/widgets/src/index.tsx
@@ -25,11 +25,14 @@ import * as mediaRequestsStats from "./media-requests/stats";
 import * as mediaServer from "./media-server";
 import * as mediaTranscoding from "./media-transcoding";
 import * as minecraftServerStatus from "./minecraft/server-status";
+import * as networkControllerStatus from "./network-controller/network-status";
+import * as networkControllerSummary from "./network-controller/summary";
 import * as notebook from "./notebook";
 import type { WidgetOptionDefinition } from "./options";
 import * as rssFeed from "./rssFeed";
 import * as smartHomeEntityState from "./smart-home/entity-state";
 import * as smartHomeExecuteAutomation from "./smart-home/execute-automation";
+import * as stockPrice from "./stocks";
 import * as video from "./video";
 import * as weather from "./weather";
 
@@ -47,11 +50,14 @@ export const widgetImports = {
   dnsHoleControls,
   "smartHome-entityState": smartHomeEntityState,
   "smartHome-executeAutomation": smartHomeExecuteAutomation,
+  stockPrice,
   mediaServer,
   calendar,
   downloads,
   "mediaRequests-requestList": mediaRequestsList,
   "mediaRequests-requestStats": mediaRequestsStats,
+  networkControllerSummary,
+  networkControllerStatus,
   rssFeed,
   bookmarks,
   indexerManager,